Remove quack.
